Fix CUDA training artifact publishing
